changed the user generated the challenges, added photo of the commentor and made some pages in the website prettier
