+++ .env
@@ -1,1 +1,1 @@
-VITE_SENDER_ID=67b9e54927d5fd857abb3637
+VITE_SENDER_ID=67c213afd5d39a96c584fd71
