+++ src/components/AI.tsx
@@ -24,32 +24,18 @@ const Chat: React.FC = () => {
     <div className="container my-4">
       <div className="d-flex justify-content-center mb-4">
         <h2
-          className="p-3"
+          className="p-2"
           style={{
-            backgroundColor: "#007bff", // צבע הרקע
+            backgroundColor: "#469187", // צבע הרקע
             color: "white", // צבע הטקסט
             padding: "10px 20px", // ריווח פנימי
             borderRadius: "5px", // עיגול קצוות
+            textAlign: "center"
           }}
         >
-          Talk With Me
+          Talk With chat <br></br>
+          Ask whatever you want about fitness, sport and eating habits
         </h2>
-      </div>
-      <div className="chat-box bg-light p-3 border rounded-3" style={{ maxHeight: "400px", overflowY: "auto" }}>
-        {messages.map((msg, index) => (
-          <div
-            key={index}
-            className={`d-flex ${msg.role === "user" ? "justify-content-end" : "justify-content-start"} mb-3`}
-          >
-            <div
-              className={`p-3 rounded-3 ${msg.role === "user" ? "bg-primary text-white" : "bg-secondary text-white"}`}
-              style={{ maxWidth: "70%", wordBreak: "break-word" }}
-            >
-              <strong>{msg.role === "user" ? "You" : "AI"}:</strong> {msg.content}
-            </div>
-          </div>
-        ))}
-        {loading && <div className="text-center text-muted">Loading...</div>}
       </div>
       <div className="d-flex mt-3">
         <input
@@ -62,6 +48,23 @@ const Chat: React.FC = () => {
         <button onClick={handleSendMessage} disabled={loading} className="btn btn-success">
           Send
         </button>
+      </div>
+      <br></br>
+      <div className="chat-box bg-light p-3 border rounded-3" style={{ maxHeight: "400px", overflowY: "auto" }}>
+        {messages.map((msg, index) => (
+          <div
+            key={index}
+            className={`d-flex ${msg.role === "user" ? "justify-content-end" : "justify-content-start"} mb-3`}
+          >
+            <div
+              className={`p-3 rounded-3 ${msg.role === "user" ? "bg-primary text-white" : "bg-secondary text-white"}`}
+              style={{ maxWidth: "70%", wordBreak: "break-word" }}
+            >
+              <strong>{msg.role === "user" ? "You" : "Chat"}:</strong> {msg.content}
+            </div>
+          </div>
+        ))}
+        {loading && <div className="text-center text-muted">Loading...</div>}
       </div>
     </div>
   );
