+++ src/components/RegistrationForm.tsx
@@ -7,7 +7,6 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faImage } from "@fortawesome/free-solid-svg-icons";
 import 'bootstrap/dist/css/bootstrap.min.css';
 import imageService from "../services/image_service";
-import { CredentialResponse, GoogleLogin } from "@react-oauth/google";
 
 interface FormData {
   username: string;
@@ -64,20 +63,6 @@ const RegistrationForm: FC = () => {
 
   const { ref, ...restRegisterParams } = register("img");
 
-  const onGoogleLoginSuccess = async (credentialResponse: CredentialResponse) => {
-    console.log(credentialResponse);
-    try{
-    const res = await userService.googleSignIn(credentialResponse);
-    console.log(res);
-    navigate('/home');
-    }catch(error){
-      console.log(error);
-    }
-  }
-
-  const onGoogleLoginError = async () => {
-    console.log("Google login failure");
-  }
   return (
     <div className="d-flex justify-content-center align-items-center vh-100">
       <div className="card p-4 shadow-sm" style={{ width: '350px' }}>
@@ -128,8 +113,10 @@ const RegistrationForm: FC = () => {
             <div className="text-danger text-center mb-3">{errorMessage}</div>
           )}
           <button type="submit" className="btn btn-primary w-100">Register</button>
-          <GoogleLogin onSuccess={onGoogleLoginSuccess} onError={onGoogleLoginError} />
         </form>
+        <div className="text-center mt-3">
+          <p className="text-muted">Already have an account? <a href="#" className="text-dark" onClick={() => navigate('/login')}>Sign In</a></p>
+        </div>
       </div>
     </div>
   );
